+++ src/common/DeepPartial.ts
@@ -1,0 +1,3 @@
+export type DeepPartial<T> = {
+    readonly [P in keyof T]?: DeepPartial<T[P]>;
+};