+++ src/entity-manager/BaseEntityManager.ts
@@ -11,6 +11,7 @@ import {NoNeedToReleaseEntityManagerError} from "./error/NoNeedToReleaseEntityMa
 import {QueryRunnerProviderAlreadyReleasedError} from "../query-runner/error/QueryRunnerProviderAlreadyReleasedError";
 import {SpecificRepository} from "../repository/SpecificRepository";
 import {MongoRepository} from "../repository/MongoRepository";
+import {DeepPartial} from "../common/DeepPartial";
 
 /**
  * Common functions shared between different entity manager types.
@@ -205,19 +206,19 @@ export abstract class BaseEntityManager {
      * Creates a new entity instance and copies all entity properties from this object into a new entity.
      * Note that it copies only properties that present in entity schema.
      */
-    create<Entity>(entityClass: ObjectType<Entity>, plainObject: Object): Entity;
+    create<Entity>(entityClass: ObjectType<Entity>, plainObject: DeepPartial<Entity>): Entity;
 
     /**
      * Creates a new entities and copies all entity properties from given objects into their new entities.
      * Note that it copies only properties that present in entity schema.
      */
-    create<Entity>(entityClass: ObjectType<Entity>, plainObjects: Object[]): Entity[];
+    create<Entity>(entityClass: ObjectType<Entity>, plainObjects: DeepPartial<Entity>[]): Entity[];
 
     /**
      * Creates a new entity instance or instances.
      * Can copy properties from the given object into new entities.
      */
-    create<Entity>(entityClass: ObjectType<Entity>, plainObjectOrObjects?: Object|Object[]): Entity|Entity[] {
+    create<Entity>(entityClass: ObjectType<Entity>, plainObjectOrObjects?: DeepPartial<Entity>|DeepPartial<Entity>[]): Entity|Entity[] {
         if (plainObjectOrObjects instanceof Array) {
             return this.getRepository(entityClass).create(plainObjectOrObjects);
 
@@ -235,15 +236,15 @@ export abstract class BaseEntityManager {
      * and returns this new entity. This new entity is actually a loaded from the db entity with all properties
      * replaced from the new object.
      */
-    preload<Entity>(entityClass: ObjectType<Entity>, object: Object): Promise<Entity> {
+    preload<Entity>(entityClass: ObjectType<Entity>, object: DeepPartial<Entity>): Promise<Entity> {
         return this.getRepository(entityClass).preload(object);
     }
 
     /**
      * Merges two entities into one new entity.
      */
-    merge<Entity>(entityClass: ObjectType<Entity>, ...objects: ObjectLiteral[]): Entity { // todo: throw exception ie tntity manager is released
-        return <Entity> this.getRepository(entityClass).merge(...objects);
+    merge<Entity>(entityClass: ObjectType<Entity>, mergeIntoEntity: Entity, ...objects: DeepPartial<Entity>[]): Entity { // todo: throw exception ie tntity manager is released
+        return <Entity> this.getRepository(entityClass).merge(mergeIntoEntity, ...objects);
     }
 
     /**