+++ test/functional/repository/basic-methods/repository-basic-methods.ts
@@ -8,6 +8,7 @@ import questionSchema from "./model-schema/QuestionSchema";
 import {Question} from "./model/Question";
 import {Blog} from "./entity/Blog";
 import {Category} from "./entity/Category";
+import {DeepPartial} from "../../../../src/common/DeepPartial";
 
 describe("repository > basic methods", () => {
 
@@ -224,10 +225,10 @@ describe("repository > basic methods", () => {
             // and preload it
             const plainBlogWithId = { id: 1 };
             const preloadedBlog = await blogRepository.preload(plainBlogWithId);
-            preloadedBlog.should.be.instanceOf(Blog);
-            preloadedBlog.id.should.be.equal(1);
-            preloadedBlog.title.should.be.equal("About people");
-            preloadedBlog.text.should.be.equal("Blog about good people");
+            preloadedBlog!.should.be.instanceOf(Blog);
+            preloadedBlog!.id.should.be.equal(1);
+            preloadedBlog!.title.should.be.equal("About people");
+            preloadedBlog!.text.should.be.equal("Blog about good people");
         })));
 
         it("should preload entity and all relations given in the object", () => Promise.all(connections.map(async connection => {
@@ -249,12 +250,12 @@ describe("repository > basic methods", () => {
             // and preload it
             const plainBlogWithId = { id: 1, categories: [{ id: 1 }] };
             const preloadedBlog = await blogRepository.preload(plainBlogWithId);
-            preloadedBlog.should.be.instanceOf(Blog);
-            preloadedBlog.id.should.be.equal(1);
-            preloadedBlog.title.should.be.equal("About people");
-            preloadedBlog.text.should.be.equal("Blog about good people");
-            preloadedBlog.categories[0].id.should.be.equal(1);
-            preloadedBlog.categories[0].name.should.be.equal("people");
+            preloadedBlog!.should.be.instanceOf(Blog);
+            preloadedBlog!.id.should.be.equal(1);
+            preloadedBlog!.title.should.be.equal("About people");
+            preloadedBlog!.text.should.be.equal("Blog about good people");
+            preloadedBlog!.categories[0].id.should.be.equal(1);
+            preloadedBlog!.categories[0].name.should.be.equal("people");
         })));
 
     });
@@ -263,6 +264,8 @@ describe("repository > basic methods", () => {
 
         it("should merge multiple entities", () => Promise.all(connections.map(async connection => {
             const blogRepository = connection.getRepository(Blog);
+
+            const originalEntity = new Blog();
 
             // first entity
             const blog1 = new Blog();
@@ -278,9 +281,10 @@ describe("repository > basic methods", () => {
             const blog3 = new Blog();
             blog3.categories = [category];
 
-            const mergedBlog = blogRepository.merge(blog1, blog2, blog3);
+            const mergedBlog = blogRepository.merge(originalEntity, blog1, blog2, blog3);
 
             mergedBlog.should.be.instanceOf(Blog);
+            mergedBlog.should.be.equal(originalEntity);
             mergedBlog.should.not.be.equal(blog1);
             mergedBlog.should.not.be.equal(blog2);
             mergedBlog.should.not.be.equal(blog3);
@@ -292,6 +296,8 @@ describe("repository > basic methods", () => {
         it("should merge both entities and plain objects", () => Promise.all(connections.map(async connection => {
             const blogRepository = connection.getRepository(Blog);
 
+            const originalEntity = new Blog();
+
             // first entity
             const blog1 = { title: "First Blog" };
 
@@ -302,9 +308,10 @@ describe("repository > basic methods", () => {
             const blog3 = new Blog();
             blog3.categories = [{ name: "category from third blog" } as Category];
 
-            const mergedBlog = blogRepository.merge(blog1, blog2, blog3);
+            const mergedBlog = blogRepository.merge(originalEntity, blog1, blog2, blog3);
 
             mergedBlog.should.be.instanceOf(Blog);
+            mergedBlog.should.be.equal(originalEntity);
             mergedBlog.should.not.be.equal(blog1);
             mergedBlog.should.not.be.equal(blog2);
             mergedBlog.should.not.be.equal(blog3);
@@ -315,7 +322,7 @@ describe("repository > basic methods", () => {
 
     });
 
-    describe("using preload and merge together", function() {
+    describe("preload also should also implement merge functionality", function() {
 
         it("if we preload entity from the plain object and merge preloaded object with plain object we'll have an object from the db with the replaced properties by a plain object's properties", () => Promise.all(connections.map(async connection => {
             const blogRepository = connection.getRepository(Blog);
@@ -339,18 +346,20 @@ describe("repository > basic methods", () => {
             await blogRepository.persist(blog);
 
             // and preload it
-            const plainBlogWithId = { id: 1, title: "changed title about people", categories: [ { id: 1 }, { id: 2, name: "insects" } ] };
+            const plainBlogWithId: DeepPartial<Blog> = {
+                id: 1,
+                title: "changed title about people",
+                categories: [ { id: 1 }, { id: 2, name: "insects" } ]
+            };
             const preloadedBlog = await blogRepository.preload(plainBlogWithId);
-            const mergedBlog = blogRepository.merge(preloadedBlog, plainBlogWithId);
-
-            mergedBlog.should.be.instanceOf(Blog);
-            mergedBlog.id.should.be.equal(1);
-            mergedBlog.title.should.be.equal("changed title about people");
-            mergedBlog.text.should.be.equal("Blog about good people");
-            mergedBlog.categories[0].id.should.be.equal(1);
-            mergedBlog.categories[0].name.should.be.equal("people");
-            mergedBlog.categories[1].id.should.be.equal(2);
-            mergedBlog.categories[1].name.should.be.equal("insects");
+            preloadedBlog!.should.be.instanceOf(Blog);
+            preloadedBlog!.id.should.be.equal(1);
+            preloadedBlog!.title.should.be.equal("changed title about people");
+            preloadedBlog!.text.should.be.equal("Blog about good people");
+            preloadedBlog!.categories[0].id.should.be.equal(1);
+            preloadedBlog!.categories[0].name.should.be.equal("people");
+            preloadedBlog!.categories[1].id.should.be.equal(2);
+            preloadedBlog!.categories[1].name.should.be.equal("insects");
         })));
 
     });